requirements.txt
packaging==24.2
wandb==0.19.8
diffusers==0.32.2
timm==1.0.3
transformers==4.48.1
sentencepiece==0.2.0
h5py==3.11.0
opencv-python==4.9.0.80
imgaug==0.4.0
huggingface_hub==0.29.3
numpy==1.26.4
tokenizers==0.21.1
safetensors
termcolor==2.5.0
draccus==0.10.0
bitsandbytes==0.42.0
datasets==3.3.2
jsonlines==4.0.0
pytest==8.3.5
tyro==0.9.5
websockets==14.1
peft
ninja
trl
deepspeed
accelerate
imageio[ffmpeg]
codetiming
liger-kernel
mathruler
omegaconf
pandas
pillow
pyarrow>=15.0.0
pylatexenc
ray[default]
tensordict
torchdata
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0 
hf-transfer
hydra-core
json_numpy
tensorflow==2.15.0
tensorflow_datasets==4.9.3
tensorflow_graphics==2021.12.3
uvicorn
fastapi
debugpy>=1.8.0
robosuite==1.4.0
bddl==1.0.1
robomimic==0.2.0 
thop==0.1.1-2209072238
future==0.18.2 
matplotlib==3.5.3 
cloudpickle==2.1.0
gym==0.25.2
easydict==1.9 
einops==0.8.0
swanlab
nvidia-ml-py